12: backtracking solution (too slow)
This is non-memoized backtracking solution that accumulates an
assignment of the '?' characters as it descents. It's structured as an
generator, which is nice in many way. It solves part 1, and faster
than the previous 2^n code, but it's still clearly exponential and
the 5-fold expansion in part 2 makes it fail.

I *believe* the written algorithm is correct, but the recursion is more
complicated than it seems it should be.

Also, the search for actual *assignments* that create a solution instead
of just the *count* of solutions (actually needed by the problem) is
not help. It also adds various high constant overheads, notably cloning
the assignment dict when a complete search is found, and there are
various other small optimizations that can be contemplated.

However, none of them seem likely to remove the fundamental backtracking
nature of the solution