12: introduce naive backtracker
`count_possible` is the straightforward translation of `backtrack` into
a form that counts solutions but doesn't generate them. It gets the same
counts as `backtrack` more directly. Still exponential of course.

It's still unsuitable for memoization because

- it accepts a mutable `assignments` dict
- it accepts a mutable `sizes` list

The `assignment` parameter can be removed, and the `lengths` array should
be replaced by a tuple. These will happen over the next commit, or maybe
a few commits.