16: convert to more typical bfs search
I wrote the original version of `follow_beam` recursively, which was
implicitly depth-first search. That was the wrong choice because
you can construct a beam loop in the input. So, I converted the function
into an iterative one using a deque to store the (pos, dir) nodes to
expand, and also track the visited (pos, dir) nodes. I kept the
depth-first nature by appending successor nodes to the *front* of the
queue.

With the `seen` set to do loop detection, this ordering doesn't matter
that much, but this commit changes `follow_beam` to *append* successor
nodes to the queue instead, turning it into a breadth-first search.